WIP: undo accidental move of ContentsHelper
