+++ textrepo-app/src/main/java/nl/knaw/huc/resources/rest/ContentsResource.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import nl.knaw.huc.helpers.ContentsHelper;
 import nl.knaw.huc.service.contents.ContentsService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
