+++ textrepo-app/src/main/java/nl/knaw/huc/resources/task/FindResource.java
@@ -4,9 +4,9 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import nl.knaw.huc.helpers.ContentsHelper;
 import nl.knaw.huc.resources.HeaderLink.Rel;
 import nl.knaw.huc.resources.HeaderLink.Uri;
-import nl.knaw.huc.resources.rest.ContentsHelper;
 import nl.knaw.huc.service.task.TaskBuilderFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
